windows: fix the two compile errors CI found
PrintWindow is classified under the Xps namespace in win32metadata, so
windows-sys exposes it as Win32::Storage::Xps::PrintWindow rather than under
WindowsAndMessaging where winuser.h would suggest. PW_RENDERFULLCONTENT does
stay in WindowsAndMessaging, so the import is split across both.

BI_RGB is a plain u32 in windows-sys, not a newtype, so .0 does not apply.

+++ windows/veil-windows/src/overlay.rs
@@ -162,11 +162,11 @@ unsafe extern "system" fn window_proc(
                 0,
                 0,
                 0,
-                DEFAULT_CHARSET,
+                DEFAULT_CHARSET as u32,
                 0,
                 0,
-                CLEARTYPE_QUALITY,
-                FF_DONTCARE,
+                CLEARTYPE_QUALITY as u32,
+                FF_DONTCARE as u32,
                 face.as_ptr() as PCWSTR,
             );
             let previous = SelectObject(hdc, font as *mut c_void);
